+++ src/Page.php
@@ -20,21 +20,14 @@ class Page {
 	protected $title;
 
 	/**
-	 * @var string
-	 */
-	protected $contentmodel;
-
-	/**
 	 * @param string $title
 	 * @param int $id
 	 * @param Revisions $revisions
-	 * @param string $contentmodel
 	 */
-	public function __construct( $title, $id, Revisions $revisions, $contentmodel ) {
+	public function __construct( $title, $id, Revisions $revisions ) {
 		$this->id = $id;
 		$this->revisions = $revisions;
 		$this->title = $title;
-		$this->contentmodel = $contentmodel;
 	}
 
 	/**
@@ -56,13 +49,6 @@ class Page {
 	 */
 	public function getTitle() {
 		return $this->title;
-	}
-
-	/**
-	 * @return string
-	 */
-	public function getContentmodel() {
-		return $this->contentmodel;
 	}
 
 }